requirements.txt
azure-identity
azure-monitor-query
pyyaml
msgraph-sdk
python-dotenv
pandas
